src/consumer/requirements.txt
python-dotenv==1.0.1
SQLAlchemy==2.0.30
asyncpg==0.29.0
aio-pika==9.4.1
